fix some -Wconversion warnings
This makes these source files compile clean with -Werror -Wconversion

BUG=webp:137
BUG=webp:355

+++ src/dsp/lossless_enc.c
@@ -703,7 +703,8 @@ static WEBP_INLINE uint32_t ClampedAddSubtractFull(uint32_t c0, uint32_t c1,
                                          (c1 >> 8) & 0xff,
                                          (c2 >> 8) & 0xff);
   const int b = AddSubtractComponentFull(c0 & 0xff, c1 & 0xff, c2 & 0xff);
-  return ((uint32_t)a << 24) | (r << 16) | (g << 8) | b;
+  return ((uint32_t)a << 24) | ((uint32_t)r << 16) |
+         ((uint32_t)g <<  8) | (uint32_t)b;
 }
 
 static WEBP_INLINE int AddSubtractComponentHalf(int a, int b) {
@@ -717,7 +718,8 @@ static WEBP_INLINE uint32_t ClampedAddSubtractHalf(uint32_t c0, uint32_t c1,
   const int r = AddSubtractComponentHalf((ave >> 16) & 0xff, (c2 >> 16) & 0xff);
   const int g = AddSubtractComponentHalf((ave >> 8) & 0xff, (c2 >> 8) & 0xff);
   const int b = AddSubtractComponentHalf((ave >> 0) & 0xff, (c2 >> 0) & 0xff);
-  return ((uint32_t)a << 24) | (r << 16) | (g << 8) | b;
+  return ((uint32_t)a << 24) | ((uint32_t)r << 16) |
+         ((uint32_t)g <<  8) | (uint32_t)b;
 }
 
 // gcc-4.9 on ARM generates incorrect code in Select() when Sub3() is inlined.
